Improve tests to suppress diffs in absolute paths
In this commit, the tests have been updated to a fully functional version. To support this, the go.mod and go.sum files were added. Additionally, the test has been updated to suppress diffs in absolute paths in expected output errors. Furthermore, the expected test error format was changed to include the error code on a new line, resulting in changes to all expected test results.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>