Format command as plugin
In this commit, a main file was added for the compare command under the name kubectl-cluster_compare, enabling execution of the command using kubectl cluster-compare. Additionally, a Makefile was added to allow easy running of tests and building of the command. This command can later be placed by users in a directory listed in the user’s PATH environment variable, allowing execution of the command as part of kubectl. Furthermore, a .gitignore file was added, and the README was moved to the root of the project.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>