Add option to format output as JSON/YAML
A new flag '-o' ('output') is introduced, allowing the user to choose the output format of the command. The options are json/yaml. The default behavior remains the same. The flag supports auto-completion for the output format options, implemented in the code of the command. To enable auto-completion when using the command as a plugin, an additional file 'kubectl_complete-cluster_compare' was added. The file should be present in a location in PATH along with the plugin to enable auto-completion. Also added supporting tests for JSON and YAML output modes.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>