Change command behavior to skip invalid resources in local mode
Previously, when using the command in local mode with recursive mode enabled, the
command would attempt to parse additional JSON/YAML files as resources. If any
files were invalid, an error was presented to the user, and the compare results
for other valid resources were not returned.

With this update, the command will now issue a warning for invalid resources
instead of stopping execution. It will skip these problematic files and proceed
to present the compare results for the valid resources.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>