Select the best match from templates in cases of multiple matches
In cases where the reference includes two templates with the same apiVersion-kind-
namespace-name 4-tuple, the command will now diff with the template that best 
matches the CR. The best match behavior can be configured in the diff config. The 
default behavior is to use the template with the least number of lines in the diff 
with the CR. It is also possible to disable this mechanism by specifying 'None' 
under the bestMatchType section in the user config.

This feature is implemented by adding an option to pass to the groupCorrelator a 
function that will return the best matched template. The function will be called by 
the correlator in case of multiple matches.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>