Add regex inline diff function
This update adds the option to include common inline diff
 functions in the tool.  Diff functions may simplify complex 
diff logic, which would be hard to write and read using 
regular Go templating and include a more tidy usage.

A new regex inline diff function allows checking fields
in cluster resources against a regex pattern. If a 
field matches the regex, no diffs are shown; if it doesn’t 
match, a diff will highlight the difference. To use this, 
enable `regexInline` for the field in `metadata.yaml`
 and add the regex in the template.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>