Reorganize container code and remove temp dir
It's best practice to remove a temporary directory manually, so this
commit reorganizes container.go to use structs/methods and adds
attributes to the Options struct in compare.go. This allows us to
make our code clearer and carry attributes over to Run(), so that we
can remove the necessary directories after the comparison has finished.